split too big class
